Add json reporting. (#10)
* Add JSON reporting.


* Fix for #26.


* Update java/benchmarks/src/main/java/javababushka/benchmarks/utils/Benchmarking.java



* Use `Optional`.


* Address PR feedback.


---------

+++ java/benchmarks/src/main/java/javababushka/benchmarks/utils/LatencyResults.java
@@ -3,13 +3,17 @@ package javababushka.benchmarks.utils;
 // Raw timing results in nanoseconds
 public class LatencyResults {
   public final double avgLatency;
-  public final long p50Latency;
-  public final long p90Latency;
-  public final long p99Latency;
+  public final double p50Latency;
+  public final double p90Latency;
+  public final double p99Latency;
   public final double stdDeviation;
 
   public LatencyResults(
-      double avgLatency, long p50Latency, long p90Latency, long p99Latency, double stdDeviation) {
+      double avgLatency,
+      double p50Latency,
+      double p90Latency,
+      double p99Latency,
+      double stdDeviation) {
     this.avgLatency = avgLatency;
     this.p50Latency = p50Latency;
     this.p90Latency = p90Latency;
